\name{PropertySet-class}
\alias{PropertySet-class}
\alias{PropertySet-class}
\title{PropertySet-class}
\description{The \code{PropertySet} class is a collection of properties. Useful
for grouping properties within a class, e.g., for storing the
parameters of some operation.}
\details{\code{PropertySet} object has following methods \describe{
\item{properties()}{Return the defined properties name and class.
You can also coerce the \code{PropertySet} object to a list by
calling as.list on the object}
}}
\author{Michael Lawrence, Tengfei Yin}

